Add adjacent text hydrate test
