+++ packages/sinuous/hydrate/test/hydrate.js
@@ -1,7 +1,6 @@
 import test from 'tape';
 import spy from 'ispy';
-import { normalizeSvg } from '../../test/_utils.js';
-import { h, hs, html, svg, hydrate, _ } from 'sinuous/hydrate';
+import { h, html, hydrate, _ } from 'sinuous/hydrate';
 import { observable } from 'sinuous';
 
 test('hydrate adds event listeners', function(t) {
@@ -199,69 +198,6 @@ test('hydrate works with a placeholder character', function(t) {
   t.end();
 });
 
-test('supports hydrating SVG via hyperscript', function(t) {
-  document.body.innerHTML = `<svg class="redbox" viewBox="0 0 100 100"><path d="M 8.74211 7.70899"></path></svg>`;
-
-  const delta = hs(
-    'svg',
-    { class: 'redbox', viewBox: '0 0 100 100' },
-    hs('path', { d: 'M 8.74211 7.70899' })
-  );
-
-  const svg = hydrate(delta, document.querySelector('svg'));
-
-  t.equal(
-    normalizeSvg(svg),
-    '<svg class="redbox" viewBox="0 0 100 100"><path d="M 8.74211 7.70899"></path></svg>'
-  );
-  t.end();
-});
-
-test('supports hydrating SVG', function(t) {
-  document.body.innerHTML = `<svg class="redbox" viewBox="0 0 100 100"><path d="M 8.74211 7.70899"></path></svg>`;
-
-  const delta = svg`
-    <svg class="redbox" viewBox="0 0 100 100">
-      <path d="M 8.74211 7.70899"></path>
-    </svg>
-  `;
-
-  const el = hydrate(delta, document.querySelector('svg'));
-
-  t.equal(
-    normalizeSvg(el),
-    '<svg class="redbox" viewBox="0 0 100 100"><path d="M 8.74211 7.70899"></path></svg>'
-  );
-  t.end();
-});
-
-test('can hydrate an array of svg elements', function(t) {
-  document.body.innerHTML = `<svg><circle cx="0" cy="1" r="10"></circle><circle cx="0" cy="2" r="10"></circle><circle cx="0" cy="3" r="10"></circle><rect x="0"></rect><circle cx="0" cy="1" r="10"></circle><circle cx="0" cy="2" r="10"></circle><circle cx="0" cy="3" r="10"></circle></svg>`;
-
-  const circles = observable([1, 2, 3]);
-  const delta = svg`<svg>
-    ${() => circles().map(c => svg`<circle cx="0" cy="${c}" r="10" />`)}
-    <rect x="0"></rect>
-    ${() => circles().map(c => svg`<circle cx="0" cy="${c}" r="10" />`)}
-  </svg>`;
-
-  const el = hydrate(delta, document.querySelector('svg'));
-
-  t.equal(
-    normalizeSvg(el),
-    '<svg><circle cx="0" cy="1" r="10"></circle><circle cx="0" cy="2" r="10"></circle><circle cx="0" cy="3" r="10"></circle><rect x="0"></rect><circle cx="0" cy="1" r="10"></circle><circle cx="0" cy="2" r="10"></circle><circle cx="0" cy="3" r="10"></circle></svg>'
-  );
-
-  circles([1, 2, 3, 4]);
-
-  t.equal(
-    normalizeSvg(el),
-    '<svg><circle cx="0" cy="1" r="10"></circle><circle cx="0" cy="2" r="10"></circle><circle cx="0" cy="3" r="10"></circle><circle cx="0" cy="4" r="10"></circle><rect x="0"></rect><circle cx="0" cy="1" r="10"></circle><circle cx="0" cy="2" r="10"></circle><circle cx="0" cy="3" r="10"></circle><circle cx="0" cy="4" r="10"></circle></svg>'
-  );
-
-  t.end();
-});
-
 test('hydrate can add a node from function', function(t) {
   document.body.innerHTML = `
     <div>
@@ -377,6 +313,27 @@ test('hydrate can add conditional observables in content', function(t) {
     div.outerHTML,
     `<div class="hamburger">Pickle Mayo Cheese
       Ham</div>`
+  );
+
+  div.parentNode.removeChild(div);
+  t.end();
+});
+
+test('hydrates adjacent text nodes', function(t) {
+  document.body.innerHTML = `
+    <div>Hi John Snow<span>!</span></div>
+  `;
+
+  const greeting = observable('Hi');
+  const name = observable('John Snow');
+  const delta = html`
+    <div>${greeting} ${name}<span>!</span></div>
+  `;
+  const div = hydrate(delta, document.querySelector('div'));
+
+  t.equal(
+    div.outerHTML,
+    `<div>Hi John Snow<span>!</span></div>`
   );
 
   div.parentNode.removeChild(div);
