+++ packages/sinuous/test/test.js
@@ -24,3 +24,4 @@ import '../map/mini/test/mini.js';
 import '../map/mini/test/mini-basic.js';
 import '../template/test/template.js';
 import '../hydrate/test/hydrate.js';
+import '../hydrate/test/svg.js';
